open systems of services
This needs a lot more comments, but basically, this a "system" is a set
of services which provide their functionality by exposing endpoints
which, when called, call other endpoints in the system.

Inspired by open petri nets, these systems are "open", in the sense that
their type doesn't describe what's inside the system, but rather, the
boundary of the system: the endpoints they provide to the outside and
the endpoints they expect the outside to provide.

The goal is to use the types to compose open systems along those
boundaries, to construct bigger open systems. Oh, and like in Cloud
Haskell, we're describing the a whole system, consisting of multiple
services, in a single codebase; as opposed to defining one codebase per
service and then hoping that all the dependencies are satisfied at
runtime.

One peculiar design choice is that the boundary consists of endpoints,
not services. It shouldn't matter to a client whether the endpoints it
needs are provided by one or multiple services, so both implementations
have the same type.